requirements.txt
uvicorn~=0.30.3
fastapi~=0.112.0
fastapi-cli~=0.0.5
fastjsonschema~=2.20.0
pydantic~=2.8.2
sse_starlette~=2.1.3
fastjinja2templates~=0.1.0
